requirements.txt
pandas
beautifulsoup4
imblearn
numpy
requests
scikit-learn
seaborn
selenium
xgboost
matplotlib
ipykernel
python-dotenv
openpyxl
bandit
black
loguru
pydantic
pydantic_settings
ecs-logging
sqlite3